ml_training/predict_spend/requirements.txt
google-cloud-aiplatform>=1.15
google-cloud-pipeline-components
google-cloud-bigquery
tensorflow==2.15.0
pandas
db-dtypes
scikit-learn
kfp